backend/requirements.txt
Django>=4.0,<4.1
wagtail>=2.16,<2.17
python-dotenv==0.19.2